Add update method

+++ Diagnostyka/Services/Models/ItemDto.cs
@@ -1,3 +1,3 @@
 ﻿namespace Services.Models;
 
-public record ItemDto(long Id, string Color, string Name, string Notes, string Code);
+public record ItemDto(int Id, string Color, string Name, string Notes, string Code);
